fix no results scrolling
