+++ frontend/packages/data-portal/app/components/TablePageLayout/index.ts
@@ -1,3 +1,3 @@
-export * from './TableHeader'
+export * from './components/components/TableHeader'
 export * from './TableHeaderDefinition'
 export * from './TablePageLayout'
